move process_masked_raster to crop.py
